ingestion_pg/pyproject.toml
[project]
name = "finance-ingestion"
version = "2.0.0"
description = "Indonesia Economy Dashboard - Data Ingestion"
requires-python = ">=3.11"
dependencies = [
    "yfinance>=0.2.36",
    "pandas>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
ingest-idx = "finance_ingestion.ingest_idx:main"
ingest-global = "finance_ingestion.ingest_global:main"
ingest-daily = "finance_ingestion.ingest_daily:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
